Fixed Database implementation and abandon MongoDb community API (unstable)
