+++ .idea/workspace.xml
@@ -1,7 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ChangeListManager">
-    <list default="true" id="e6b441f1-f6e5-4fb3-8dc4-feb8c1acbd14" name="Changes" comment="" />
+    <list default="true" id="e6b441f1-f6e5-4fb3-8dc4-feb8c1acbd14" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Box-Office-Mojo-API-master/.vscode/settings.json" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Box-Office-Mojo-API-master/README.md" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Box-Office-Mojo-API-master/boxofficemojo.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Box-Office-Mojo-API-master/movie.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/DAL/DatabaseHelper.py" beforeDir="false" afterPath="$PROJECT_DIR$/DAL/DatabaseHelper.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/DAL/Movie/MovieDal.py" beforeDir="false" afterPath="$PROJECT_DIR$/DAL/Movie_Entity/MovieDal.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/DAL/Movie/MovieData.py" beforeDir="false" afterPath="$PROJECT_DIR$/DAL/Movie_Entity/MovieData.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/DAL/Tweet/TweetDal.py" beforeDir="false" afterPath="$PROJECT_DIR$/DAL/Tweet_Entity/TweetDal.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/DAL/Tweet/TweetData.py" beforeDir="false" afterPath="$PROJECT_DIR$/DAL/Tweet_Entity/TweetData.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/main.py" afterDir="false" />
+    </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -14,25 +26,55 @@
       </list>
     </option>
   </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
   <component name="MarkdownSettingsMigration">
     <option name="stateVersion" value="1" />
   </component>
   <component name="ProjectId" id="2CzNHdfn25xbzWUtwg9nDfRFl9X" />
+  <component name="ProjectLevelVcsManager" settingsEditedManually="true">
+    <ConfirmationsSetting value="2" id="Add" />
+  </component>
   <component name="ProjectViewState">
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
   <component name="PropertiesComponent">{
   &quot;keyToString&quot;: {
-    &quot;settings.editor.selected.configurable&quot;: &quot;org.jetbrains.plugins.github.ui.GithubSettingsConfigurable&quot;
+    &quot;ASKED_ADD_EXTERNAL_FILES&quot;: &quot;true&quot;,
+    &quot;settings.editor.selected.configurable&quot;: &quot;com.jetbrains.python.configuration.PyActiveSdkModuleConfigurable&quot;
   }
 }</component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="E:\AFEKA\Finals Porjects\Code\DAL" />
+      <recent name="E:\AFEKA\Finals Porjects\Code\Box-Office-Mojo-API" />
       <recent name="E:\AFEKA\Finals Porjects\Code\DAL\Movie" />
     </key>
   </component>
-  <component name="RunManager">
+  <component name="RunManager" selected="Python.main">
+    <configuration name="TweetData" type="PythonConfigurationType" factoryName="Python" temporary="true" nameIsGenerated="true">
+      <module name="Code" />
+      <option name="INTERPRETER_OPTIONS" value="" />
+      <option name="PARENT_ENVS" value="true" />
+      <envs>
+        <env name="PYTHONUNBUFFERED" value="1" />
+      </envs>
+      <option name="SDK_HOME" value="" />
+      <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/DAL/Tweet_Entity" />
+      <option name="IS_MODULE_SDK" value="true" />
+      <option name="ADD_CONTENT_ROOTS" value="true" />
+      <option name="ADD_SOURCE_ROOTS" value="true" />
+      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/DAL/Tweet_Entity/TweetData.py" />
+      <option name="PARAMETERS" value="" />
+      <option name="SHOW_COMMAND_LINE" value="false" />
+      <option name="EMULATE_TERMINAL" value="false" />
+      <option name="MODULE_MODE" value="false" />
+      <option name="REDIRECT_INPUT" value="false" />
+      <option name="INPUT_FILE" value="" />
+      <method v="2" />
+    </configuration>
     <configuration name="main" type="PythonConfigurationType" factoryName="Python" nameIsGenerated="true">
       <module name="Code" />
       <option name="INTERPRETER_OPTIONS" value="" />
@@ -54,6 +96,11 @@
       <option name="INPUT_FILE" value="" />
       <method v="2" />
     </configuration>
+    <recent_temporary>
+      <list>
+        <item itemvalue="Python.TweetData" />
+      </list>
+    </recent_temporary>
   </component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
   <component name="TaskManager">
@@ -65,5 +112,27 @@
       <updated>1659801366325</updated>
     </task>
     <servers />
+  </component>
+  <component name="Vcs.Log.Tabs.Properties">
+    <option name="TAB_STATES">
+      <map>
+        <entry key="MAIN">
+          <value>
+            <State />
+          </value>
+        </entry>
+      </map>
+    </option>
+  </component>
+  <component name="XDebuggerManager">
+    <breakpoint-manager>
+      <breakpoints>
+        <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
+          <url>file://$PROJECT_DIR$/main.py</url>
+          <line>17</line>
+          <option name="timeStamp" value="1" />
+        </line-breakpoint>
+      </breakpoints>
+    </breakpoint-manager>
   </component>
 </project>
