--- Box-Office-Mojo-API-master/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "/usr/local/bin/python3"
-}
